chore(cli): update shopify template to use strict mode schemas
