+++ packages/@sanity/cli/templates/shopify/schemas/blocks/body.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import {defineField} from 'sanity'
 
 export default defineField({
@@ -50,27 +49,21 @@ export default defineField({
     },
     // Custom blocks
     {
-      name: 'blockAccordion',
       type: 'module.accordion',
     },
     {
-      name: 'blockCallout',
       type: 'module.callout',
     },
     {
-      name: 'blockGrid',
       type: 'module.grid',
     },
     {
-      name: 'blockImages',
       type: 'module.images',
     },
     {
-      name: 'blockInstagram',
       type: 'module.instagram',
     },
     {
-      name: 'blockProducts',
       type: 'module.products',
     },
   ],
