+++ packages/@sanity/cli/templates/shopify/schemas/documents/colorTheme.tsx
@@ -1,6 +1,7 @@
 import {IceCreamIcon} from '@sanity/icons'
 import React from 'react'
 import {defineField, defineType} from 'sanity'
+
 import ColorTheme from '../../components/media/ColorTheme'
 
 export default defineType({
