+++ packages/@sanity/cli/templates/shopify/schemas/documents/product.tsx
@@ -10,6 +10,7 @@ const GROUPS = [
   {
     name: 'editorial',
     title: 'Editorial',
+    default: true,
   },
   {
     name: 'shopifySync',
