+++ packages/@sanity/cli/templates/shopify/schemas/index.ts
@@ -33,67 +33,103 @@ const blocks = [body]
 
 // Object types
 import customProductOptionColor from './objects/customProductOption/color'
+import customProductOptionColorObject from './objects/customProductOption/colorObject'
 import customProductOptionSize from './objects/customProductOption/size'
-import imageWithProductHotspots from './objects/imageWithProductHotspots'
-import linkExternal from './objects/linkExternal'
-import linkInternal from './objects/linkInternal'
+import customProductOptionSizeObject from './objects/customProductOption/sizeObject'
+import footer from './objects/global/footer'
+import imageWithProductHotspots from './objects/hotspot/imageWithProductHotspots'
+import inventory from './objects/shopify/inventory'
+import linkExternal from './objects/global/linkExternal'
+import linkInternal from './objects/global/linkInternal'
+import links from './objects/global/links'
+import notFoundPage from './objects/global/notFoundPage'
 import heroCollection from './objects/hero/collection'
 import heroHome from './objects/hero/home'
 import heroPage from './objects/hero/page'
 import moduleAccordion from './objects/module/accordion'
+import accordionBody from './objects/module/accordionBody'
+import accordionGroup from './objects/module/accordionGroup'
 import moduleCallout from './objects/module/callout'
 import moduleCallToAction from './objects/module/callToAction'
 import moduleCollection from './objects/module/collection'
 import moduleGrid from './objects/module/grid'
+import gridItems from './objects/module/gridItem'
+import menu from './objects/global/menu'
 import moduleImage from './objects/module/image'
+import moduleImageAction from './objects/module/imageCallToAction'
 import moduleImages from './objects/module/images'
 import moduleInstagram from './objects/module/instagram'
 import moduleProduct from './objects/module/product'
 import moduleProducts from './objects/module/products'
-import placeholderString from './objects/placeholderString'
-import productHotspots from './objects/productHotspots'
-import productOption from './objects/productOption'
-import productWithVariant from './objects/productWithVariant'
-import proxyString from './objects/proxyString'
+import placeholderString from './objects/shopify/placeholderString'
+import priceRange from './objects/shopify/priceRange'
+import spot from './objects/hotspot/spot'
+import productHotspots from './objects/hotspot/productHotspots'
+import option from './objects/shopify/option'
+import productWithVariant from './objects/shopify/productWithVariant'
+import proxyString from './objects/shopify/proxyString'
+import seo from './objects/seo/seo'
 import seoHome from './objects/seo/home'
 import seoPage from './objects/seo/page'
+import seoDescription from './objects/seo/description'
 import seoShopify from './objects/seo/shopify'
-import shopifyCollection from './objects/shopifyCollection'
-import shopifyCollectionRule from './objects/shopifyCollectionRule'
-import shopifyProduct from './objects/shopifyProduct'
-import shopifyProductVariant from './objects/shopifyProductVariant'
+import shopifyCollection from './objects/shopify/shopifyCollection'
+import shopifyCollectionRule from './objects/shopify/shopifyCollectionRule'
+import shopifyProduct from './objects/shopify/shopifyProduct'
+import shopifyProductVariant from './objects/shopify/shopifyProductVariant'
+
+// Collections
+import collectionGroup from './objects/collection/group'
+import collectionLinks from './objects/collection/links'
 
 const objects = [
   customProductOptionColor,
+  customProductOptionColorObject,
   customProductOptionSize,
+  customProductOptionSizeObject,
+  footer,
   imageWithProductHotspots,
+  inventory,
+  links,
   linkExternal,
   linkInternal,
+  notFoundPage,
   heroCollection,
   heroHome,
   heroPage,
   moduleAccordion,
+  accordionBody,
+  accordionGroup,
+  menu,
   moduleCallout,
   moduleCallToAction,
   moduleCollection,
   moduleGrid,
+  gridItems,
   moduleImage,
+  moduleImageAction,
   moduleImages,
   moduleInstagram,
   moduleProduct,
   moduleProducts,
   placeholderString,
+  priceRange,
+  spot,
   productHotspots,
-  productOption,
+  option,
   productWithVariant,
   proxyString,
+  seo,
   seoHome,
   seoPage,
+  seoDescription,
   seoShopify,
   shopifyCollection,
   shopifyCollectionRule,
   shopifyProduct,
   shopifyProductVariant,
+  collectionGroup,
+  collectionLinks,
 ]
 
-export const schemaTypes = [...annotations, ...documents, ...singletons, ...objects, ...blocks]
+export const schemaTypes = [...annotations, ...singletons, ...objects, ...blocks, ...documents]
