+++ packages/@sanity/cli/templates/shopify/schemas/objects/customProductOption/size.ts
@@ -1,5 +1,6 @@
 import pluralize from 'pluralize-esm'
 import {defineField} from 'sanity'
+
 interface SizeOption {
   title: string
 }
@@ -25,46 +26,7 @@ export default defineField({
       type: 'array',
       of: [
         {
-          name: 'size',
-          title: 'Size',
-          type: 'object',
-          fields: [
-            defineField({
-              name: 'title',
-              title: 'Title',
-              type: 'string',
-              description: 'Shopify product option value (case sensitive)',
-              validation: (Rule) => Rule.required(),
-            }),
-            defineField({
-              name: 'width',
-              title: 'Width',
-              type: 'number',
-              description: 'In mm',
-              validation: (Rule) => Rule.required().precision(2),
-            }),
-            defineField({
-              name: 'height',
-              title: 'Height',
-              type: 'number',
-              description: 'In mm',
-              validation: (Rule) => Rule.required().precision(2),
-            }),
-          ],
-          preview: {
-            select: {
-              height: 'height',
-              title: 'title',
-              width: 'width',
-            },
-            prepare(selection) {
-              const {height, title, width} = selection
-              return {
-                subtitle: `${width || '??'}cm x ${height || '??'}cm`,
-                title,
-              }
-            },
-          },
+          type: 'customProductOption.sizeObject',
         },
       ],
       validation: (Rule) =>
