+++ packages/@sanity/cli/templates/shopify/schemas/objects/global/linkExternal.ts
