+++ packages/@sanity/cli/templates/shopify/schemas/objects/global/linkInternal.ts
@@ -1,7 +1,8 @@
 import {LinkIcon} from '@sanity/icons'
 import {defineField} from 'sanity'
-import {PAGE_REFERENCES} from '../../constants'
-import {getPriceRange} from '../../utils/getPriceRange'
+
+import {PAGE_REFERENCES} from '../../../constants'
+import {getPriceRange} from '../../../utils/getPriceRange'
 
 export default defineField({
   title: 'Internal Link',
