+++ packages/@sanity/cli/templates/shopify/schemas/objects/global/links.ts
@@ -1,0 +1,16 @@
+import {defineField} from 'sanity'
+
+export default defineField({
+  name: 'menuLinks',
+  title: 'menuLinks',
+  type: 'array',
+  of: [
+    {
+      name: 'collectionGroup',
+      title: 'Collection group',
+      type: 'collectionGroup',
+    },
+    {type: 'linkInternal'},
+    {type: 'linkExternal'},
+  ],
+})
