+++ packages/@sanity/cli/templates/shopify/schemas/objects/global/menu.ts
@@ -1,0 +1,19 @@
+import {defineField} from 'sanity'
+
+export default defineField({
+  name: 'menuSettings',
+  title: 'Menu',
+  type: 'object',
+  options: {
+    collapsed: false,
+    collapsible: true,
+  },
+  fields: [
+    // Links
+    defineField({
+      name: 'links',
+      title: 'Links',
+      type: 'menuLinks',
+    }),
+  ],
+})
