+++ packages/@sanity/cli/templates/shopify/schemas/objects/hotspot/imageWithProductHotspots.ts
