+++ packages/@sanity/cli/templates/shopify/schemas/objects/module/grid.ts
@@ -1,7 +1,6 @@
 import {ThLargeIcon} from '@sanity/icons'
 import pluralize from 'pluralize-esm'
 import {defineField} from 'sanity'
-import blocksToText from '../../../utils/blocksToText'
 
 export default defineField({
   name: 'module.grid',
@@ -16,91 +15,7 @@ export default defineField({
       type: 'array',
       of: [
         {
-          name: 'item',
-          title: 'Item',
-          type: 'object',
-          fields: [
-            // Title
-            defineField({
-              name: 'title',
-              title: 'Title',
-              type: 'string',
-              validation: (Rule) => Rule.required(),
-            }),
-            // Image
-            defineField({
-              name: 'image',
-              title: 'Image',
-              type: 'image',
-              options: {hotspot: true},
-              validation: (Rule) => Rule.required(),
-            }),
-            // Body
-            defineField({
-              name: 'body',
-              title: 'Body',
-              type: 'array',
-              of: [
-                {
-                  lists: [],
-                  marks: {
-                    annotations: [
-                      // Product
-                      {
-                        name: 'annotationProduct',
-                        type: 'annotationProduct',
-                      },
-                      // Email
-                      {
-                        name: 'annotationLinkEmail',
-                        type: 'annotationLinkEmail',
-                      },
-                      // Internal link
-                      {
-                        name: 'annotationLinkInternal',
-                        type: 'annotationLinkInternal',
-                      },
-                      // URL
-                      {
-                        name: 'annotationLinkExternal',
-                        type: 'annotationLinkExternal',
-                      },
-                    ],
-                    decorators: [
-                      {
-                        title: 'Italic',
-                        value: 'em',
-                      },
-                      {
-                        title: 'Strong',
-                        value: 'strong',
-                      },
-                    ],
-                  },
-                  // Regular styles
-                  styles: [],
-                  // Paragraphs
-                  type: 'block',
-                },
-              ],
-              validation: (Rule) => Rule.required(),
-            }),
-          ],
-          preview: {
-            select: {
-              body: 'body',
-              image: 'image',
-              title: 'title',
-            },
-            prepare(selection) {
-              const {body, image, title} = selection
-              return {
-                media: image,
-                subtitle: body && blocksToText(body),
-                title,
-              }
-            },
-          },
+          type: 'gridItem',
         },
       ],
     },
@@ -114,7 +29,7 @@ export default defineField({
       const {items} = selection
       return {
         subtitle: 'Grid',
-        title: items.length > 0 ? pluralize('item', items.length, true) : 'No items',
+        title: items?.length > 0 ? pluralize('item', items.length, true) : 'No items',
       }
     },
   },
