+++ packages/@sanity/cli/templates/shopify/schemas/objects/module/imageCallToAction.tsx
@@ -1,0 +1,23 @@
+import {defineField} from 'sanity'
+
+export default defineField({
+  name: 'imageCallToAction',
+  title: 'Call to action',
+  type: 'object',
+  fields: [
+    // Title
+    {
+      name: 'title',
+      title: 'Title',
+      type: 'string',
+    },
+    // Link
+    {
+      name: 'links',
+      title: 'Link',
+      type: 'array',
+      of: [{type: 'linkInternal'}, {type: 'linkExternal'}],
+      validation: (Rule) => Rule.max(1),
+    },
+  ],
+})
