+++ packages/@sanity/cli/templates/shopify/schemas/objects/module/images.tsx
@@ -61,7 +61,7 @@ export default defineField({
       const {images} = selection
       return {
         subtitle: 'Images',
-        title: images.length > 0 ? pluralize('image', images.length, true) : 'No images',
+        title: images?.length > 0 ? pluralize('image', images.length, true) : 'No images',
         media: ImageIcon,
       }
     },
