+++ packages/@sanity/cli/templates/shopify/schemas/objects/seo/description.tsx
@@ -1,0 +1,10 @@
+import {defineField} from 'sanity'
+
+export default defineField({
+  name: 'seo.description',
+  title: 'Description',
+  type: 'text',
+  rows: 3,
+  validation: (Rule) =>
+    Rule.max(150).warning('Longer descriptions may be truncated by search engines'),
+})
