+++ packages/@sanity/cli/templates/shopify/schemas/objects/seo/shopify.tsx
@@ -29,10 +29,7 @@ export default defineField({
     {
       name: 'description',
       title: 'Description',
-      type: 'text',
-      rows: 2,
-      validation: (Rule) =>
-        Rule.max(150).warning('Longer descriptions may be truncated by search engines'),
+      type: 'seo.description',
     },
     {
       name: 'image',
