+++ packages/@sanity/cli/templates/shopify/schemas/objects/shopify/inventory.ts
@@ -1,0 +1,30 @@
+import {defineField} from 'sanity'
+
+export default defineField({
+  name: 'inventory',
+  title: 'Inventory',
+  type: 'object',
+  options: {
+    columns: 3,
+  },
+  fields: [
+    // Available
+    {
+      name: 'isAvailable',
+      title: 'Available',
+      type: 'boolean',
+    },
+    // Management
+    {
+      name: 'management',
+      title: 'Management',
+      type: 'string',
+    },
+    // Policy
+    {
+      name: 'policy',
+      title: 'Policy',
+      type: 'string',
+    },
+  ],
+})
