+++ packages/@sanity/cli/templates/shopify/schemas/objects/shopify/productWithVariant.tsx
@@ -2,9 +2,10 @@ import {TagIcon} from '@sanity/icons'
 import pluralize from 'pluralize-esm'
 import React from 'react'
 import {defineField} from 'sanity'
-import ShopifyDocumentStatus from '../../components/media/ShopifyDocumentStatus'
-import {SANITY_API_VERSION} from '../../constants'
-import {getPriceRange} from '../../utils/getPriceRange'
+
+import ShopifyDocumentStatus from '../../../components/media/ShopifyDocumentStatus'
+import {SANITY_API_VERSION} from '../../../constants'
+import {getPriceRange} from '../../../utils/getPriceRange'
 
 export default defineField({
   name: 'productWithVariant',
