+++ packages/@sanity/cli/templates/shopify/schemas/objects/shopify/shopifyCollection.ts
@@ -85,8 +85,6 @@ export default defineField({
       description: 'Include Shopify products that satisfy these conditions',
       of: [
         {
-          name: 'rule',
-          title: 'Rule',
           type: 'collectionRule',
         },
       ],
