+++ packages/@sanity/cli/templates/shopify/schemas/objects/shopify/shopifyCollectionRule.tsx
