+++ packages/@sanity/cli/templates/shopify/schemas/objects/shopify/shopifyProduct.ts
@@ -123,22 +123,7 @@ export default defineField({
     defineField({
       name: 'priceRange',
       title: 'Price range',
-      type: 'object',
-      options: {
-        columns: 2,
-      },
-      fields: [
-        {
-          name: 'minVariantPrice',
-          title: 'Min variant price',
-          type: 'number',
-        },
-        {
-          name: 'maxVariantPrice',
-          title: 'Max variant price',
-          type: 'number',
-        },
-      ],
+      type: 'priceRange',
     }),
     // Preview Image URL
     defineField({
@@ -154,9 +139,7 @@ export default defineField({
       type: 'array',
       of: [
         {
-          name: 'option',
-          title: 'Option',
-          type: 'productOption',
+          type: 'option',
         },
       ],
     }),
