+++ packages/@sanity/cli/templates/shopify/schemas/singletons/settings.ts
@@ -1,4 +1,4 @@
-import {CogIcon, PackageIcon} from '@sanity/icons'
+import {CogIcon} from '@sanity/icons'
 import {defineType, defineField} from 'sanity'
 
 const TITLE = 'Settings'
@@ -35,117 +35,15 @@ export default defineType({
     defineField({
       name: 'menu',
       title: 'Menu',
-      type: 'object',
+      type: 'menuSettings',
       group: 'navigation',
-      options: {
-        collapsed: false,
-        collapsible: true,
-      },
-      fields: [
-        // Links
-        defineField({
-          name: 'links',
-          title: 'Links',
-          type: 'array',
-          of: [
-            {
-              name: 'collectionGroup',
-              title: 'Collection group',
-              type: 'object',
-              icon: PackageIcon,
-              fields: [
-                {
-                  name: 'title',
-                  title: 'Title',
-                  type: 'string',
-                  validation: (Rule) => Rule.required(),
-                },
-                {
-                  name: 'collectionLinks',
-                  title: 'Collection links',
-                  type: 'array',
-                  validation: (Rule) => Rule.unique().max(4),
-                  of: [
-                    {
-                      name: 'collection',
-                      type: 'reference',
-                      weak: true,
-                      to: [{type: 'collection'}],
-                    },
-                  ],
-                },
-                {
-                  name: 'collectionProducts',
-                  title: 'Collection products',
-                  type: 'reference',
-                  description: 'Products from this collection will be listed',
-                  weak: true,
-                  to: [{type: 'collection'}],
-                },
-              ],
-            },
-            {type: 'linkInternal'},
-            {type: 'linkExternal'},
-          ],
-        }),
-      ],
     }),
     // Footer
     defineField({
       name: 'footer',
       title: 'Footer',
-      type: 'object',
+      type: 'footerSettings',
       group: 'navigation',
-      options: {
-        collapsed: false,
-        collapsible: true,
-      },
-      fields: [
-        // Links
-        defineField({
-          name: 'links',
-          title: 'Links',
-          type: 'array',
-          of: [{type: 'linkInternal'}, {type: 'linkExternal'}],
-        }),
-        // Text
-        defineField({
-          name: 'text',
-          title: 'Text',
-          type: 'array',
-          of: [
-            {
-              lists: [],
-              marks: {
-                annotations: [
-                  // Email
-                  {
-                    title: 'Email',
-                    name: 'annotationLinkEmail',
-                    type: 'annotationLinkEmail',
-                  },
-                  // Internal link
-                  {
-                    title: 'Internal page',
-                    name: 'annotationLinkInternal',
-                    type: 'annotationLinkInternal',
-                  },
-                  // URL
-                  {
-                    title: 'URL',
-                    name: 'annotationLinkExternal',
-                    type: 'annotationLinkExternal',
-                  },
-                ],
-                decorators: [],
-              },
-              // Block styles
-              styles: [{title: 'Normal', value: 'normal'}],
-              type: 'block',
-            },
-          ],
-        }),
-      ],
     }),
     // Custom product options
     defineField({
@@ -179,71 +77,15 @@ export default defineType({
     defineField({
       name: 'notFoundPage',
       title: '404 page',
-      type: 'object',
+      type: 'notFoundPage',
       group: 'notFoundPage',
-      fields: [
-        defineField({
-          name: 'title',
-          title: 'Title',
-          type: 'string',
-          validation: (Rule) => Rule.required(),
-        }),
-        defineField({
-          name: 'body',
-          title: 'Body',
-          type: 'text',
-          rows: 2,
-        }),
-        defineField({
-          name: 'collection',
-          title: 'Collection',
-          type: 'reference',
-          description: 'Collection products displayed on this page',
-          weak: true,
-          to: [
-            {
-              name: 'collection',
-              type: 'collection',
-            },
-          ],
-        }),
-        // Color theme
-        defineField({
-          name: 'colorTheme',
-          title: 'Color theme',
-          type: 'reference',
-          to: [{type: 'colorTheme'}],
-        }),
-      ],
     }),
     // SEO
     defineField({
       name: 'seo',
       title: 'SEO',
-      type: 'object',
+      type: 'seo',
       group: 'seo',
-      description: 'Defaults for every page',
-      options: {
-        collapsed: false,
-        collapsible: true,
-      },
-      fields: [
-        defineField({
-          name: 'title',
-          title: 'Site title',
-          type: 'string',
-          validation: (Rule) => Rule.required(),
-        }),
-        defineField({
-          name: 'description',
-          title: 'Description',
-          type: 'text',
-          rows: 2,
-          validation: (Rule) =>
-            Rule.max(150).warning('Longer descriptions may be truncated by search engines'),
-        }),
-      ],
-      validation: (Rule) => Rule.required(),
     }),
   ],
   preview: {
